+++ build.gradle
@@ -73,6 +73,6 @@ sonarqube {
     properties {
         property "sonar.projectName", "Gradle Minify Plugin"
         property "sonar.projectKey", "616slayer616_gradle-minify-plugin"
-        property "sonar.coverage.exclusions", '**/*Extension,**/*MinifyTask'
+        property "sonar.coverage.exclusions", '**/*Extension,**/*MinifyTaskBase'
     }
 }
