+++ src/main/java/org/padler/gradle/minify/MinifyTask.java
@@ -1,70 +1,12 @@
 package org.padler.gradle.minify;
 
-import org.gradle.api.DefaultTask;
-import org.gradle.api.tasks.*;
+import org.gradle.api.tasks.TaskAction;
 import org.padler.gradle.minify.minifier.CssMinifier;
 import org.padler.gradle.minify.minifier.JsMinifier;
 import org.padler.gradle.minify.minifier.options.CSSMinifierOptions;
 import org.padler.gradle.minify.minifier.options.JSMinifierOptions;
 
-import java.io.File;
-
-public class MinifyTask extends DefaultTask {
-
-    MinifyPluginExtension extension = getProject().getExtensions().findByType(MinifyPluginExtension.class);
-
-    public MinifyTask() {
-        if (extension == null) {
-            extension = new MinifyPluginExtension();
-        }
-    }
-
-    @Optional
-    @InputDirectory
-    public File getJsSrcDir() {
-        if (extension.getJsSrcDir().isEmpty()) return null;
-        return new File(extension.getJsSrcDir());
-    }
-
-    @Optional
-    @InputDirectory
-    public File getCssSrcDir() {
-        if (extension.getCssSrcDir().isEmpty()) return null;
-        return new File(extension.getCssSrcDir());
-    }
-
-    @Optional
-    @OutputDirectory
-    public File getJsDstDir() {
-        if (extension.getJsDstDir().isEmpty()) return null;
-        return new File(extension.getJsDstDir());
-    }
-
-    @Optional
-    @OutputDirectory
-    public File getCssDstDir() {
-        if (extension.getCssDstDir().isEmpty()) return null;
-        return new File(extension.getCssDstDir());
-    }
-
-    @Optional
-    @Input
-    public Boolean getCreateCssSourceMaps() {
-        return extension.getCreateCssSourceMaps();
-    }
-
-    @Optional
-    @Input
-    public Boolean getCreateJsSourceMaps() {
-        return extension.getCreateJsSourceMaps();
-    }
-
-    @Optional
-    @Input
-    public Boolean getOriginalFileNames() {
-        return extension.getOriginalFileNames();
-    }
-
+public class MinifyTask extends MinifyTaskBase {
     @TaskAction
     public void minify() {
         if (!extension.getCssSrcDir().isEmpty() && !extension.getCssDstDir().isEmpty()) {
